+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 ### documentdb v1.109-0 (Unreleased) ###
+* Support collation with find positional queries *[Feature]*
 
 ### documentdb v0.108-0 (Unreleased) ###
 * Top-level `let` variables and `$$NOW` supported by default.